+++ src/main/java/roomescape/service/ReservationService.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Service;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationDate;
 import roomescape.domain.reservation.Waiting;
-import roomescape.domain.reservation.WaitingWithRank;
 import roomescape.domain.user.Member;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.WaitingRepository;
+import roomescape.repository.dto.WaitingWithRank;
 import roomescape.service.dto.input.ReservationInput;
 import roomescape.service.dto.input.ReservationSearchInput;
 import roomescape.service.dto.output.ReservationOutput;
